+++ gomusic_server/dto/collect_request.go
@@ -7,8 +7,8 @@ import (
 )
 
 type CollectRequest struct {
-	ID         uint8      `json:"id,omitempty"`
-	UserID     int8       `json:"userId,omitempty"`
+	ID         uint       `json:"id,omitempty"`
+	UserID     int        `json:"userId,omitempty"`
 	Type       string     `json:"type,omitempty"`
 	SongID     int        `json:"songId,omitempty"`
 	SongListID string     `json:"songListId,omitempty"`
@@ -29,11 +29,11 @@ func (req *CollectRequest) ToCollect() models.Collection {
 	songListID, _ := strconv.Atoi(req.SongListID)
 
 	return models.Collection{
-		ID:         int(req.ID),     // uint8 转换为 int
-		UserID:     int(req.UserID), // int8 转换为 int
+		ID:         int(req.ID),     // uint 转换为 int
+		UserID:     int(req.UserID), // int 转换为 int
 		Type:       songType,        // string 转换为 int
-		SongID:     req.SongID,      // int8 转换为 int
-		SongListID: songListID,      // int8 转换为 int
+		SongID:     req.SongID,      // int 转换为 int
+		SongListID: songListID,      // int 转换为 int
 		CreateTime: createTime,      // 转换后的时间
 	}
 }
