+++ gomusic_server/services/rank_list_service.go
@@ -15,12 +15,12 @@ func NewRankListService(db *gorm.DB) *RankListService {
 	}
 }
 
-func (s *RankListService) RankOfSongListId(songListId int8) (int64, float64, error) {
+func (s *RankListService) RankOfSongListId(songListId int) (int64, float64, error) {
 	count, scoreSum, err := s.rankListDAO.RankOfSongListId(songListId)
 	return count, scoreSum, err
 }
 
-func (s *RankListService) GetUserRank(consumerid int8, songListId int8) (uint32, error) {
+func (s *RankListService) GetUserRank(consumerid int, songListId int) (uint32, error) {
 	score, err := s.rankListDAO.GetUserRank(consumerid, songListId)
 	return score, err
 }
